+++ app/controllers/ServiciosController.php
@@ -66,42 +66,7 @@ class ServiciosController extends \BaseController {
 	}
 
 
-	/**
-	 * Display the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function show($id)
-	{
-		//
-	}
-
-
-	/**
-	 * Show the form for editing the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function edit($id)
-	{
-		//
-	}
-
-
-	/**
-	 * Update the specified resource in storage.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function update($id)
-	{
-		//
-	}
-
-
+	
 	/**
 	 * Remove the specified resource from storage.
 	 *
@@ -110,6 +75,15 @@ class ServiciosController extends \BaseController {
 	 */
 	public function destroyServicio($id)
 	{
+
+		$tickets = Ticket::where('servicio_id',$id)->get();
+
+		if(count($tickets) > 0 )
+		{
+			return Redirect::route('ver-servicio')
+			->with('global-error','El servicio no puede ser borrado. Hay tickets con este servicio asignado.');
+		}
+
 		Servicio::destroy($id);
 		return Redirect::route('ver-servicio');
 	}
